requirements.txt
typer[all]
httpx
lancedb
sentence-transformers
pyyaml
fastapi
uvicorn
beautifulsoup4
scikit-learn
